renamed get_column to typed_column
