+++ rust/analytics/src/dfext/mod.rs
@@ -2,8 +2,6 @@
 pub mod async_log_stream;
 /// Utilities to help deal with df expressions
 pub mod expressions;
-/// Access to a RecordBatch's columns
-pub mod get_column;
 /// Stream a function's log as a table
 pub mod log_stream_table_provider;
 /// Get min & max from the time column
@@ -12,3 +10,5 @@ pub mod min_max_time_df;
 pub mod predicate;
 /// Execution plan interface for an async task
 pub mod task_log_exec_plan;
+/// Access to a RecordBatch's columns
+pub mod typed_column;
